+++ src/private/quality/GetNovaPesterTestResultPath.ps1
@@ -1,8 +1,9 @@
 function Get-NovaPesterTestResultPath {
     [CmdletBinding()]
     param(
-        [Parameter(Mandatory)][string]$ProjectRoot
+        [Parameter(Mandatory)][string]$ProjectRoot,
+        [Parameter()][string]$FileName = 'TestResults.xml'
     )
 
-    return [System.IO.Path]::Join($ProjectRoot, 'artifacts', 'TestResults.xml')
+    return [System.IO.Path]::Join($ProjectRoot, 'artifacts', $FileName)
 }